+++ src/ogg/vorbis/write.rs
@@ -37,7 +37,7 @@ pub(crate) fn write_to(
 		let p = Page::read(data, false)?;
 
 		if p.header_type() & 0x01 != 1 {
-			data.seek(SeekFrom::Start(p.start as u64))?;
+			data.seek(SeekFrom::Start(p.start))?;
 			data.read_to_end(&mut remaining)?;
 
 			reached_md_end = true;
