+++ src/types/picture.rs
@@ -803,7 +803,7 @@ impl Picture {
 		size -= 20;
 
 		if data_len <= size {
-			let mut data = try_vec![0; data_len as usize];
+			let mut data = try_vec![0; data_len];
 
 			if let Ok(()) = reader.read_exact(&mut data) {
 				return Ok((
